requirements.txt
streamlit
faster-whisper
ctranslate2==3.20.0 --extra-index-url https://download.argosopentech.com/whl/cpu
ffmpeg-python
transformers
torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
yt-dlp
